# server/schemas/openApiInterface.yml
openapi: "3.0.0"
info:
  title: Simple API overview
  version: 2.0.0
components:
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
    JWT:     # <-- arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
paths:
...
  /users/:
    get:
      summary: Retrieves all users
      tags:
        - Users
      security:
        - JWT: []
      responses:
        200:
          description: Users retrieved
          content:
            application/json:
              schema:
                properties:
                  UserID:
                    type: integer
                  role:
                    type: string
                  Username:
                    type: string
                  UserPassword:
                    type: string
                  Created:
                    type: string
              example:
                  [
                    {
                      UserID       : 1,
                      role         : "admin",
                      Username     : "user1",
                      UserPassword : "123",
                      Created      : "2024-06-02T19:29:50.000Z"
                    }
                  ]
        401:
          description: User not authorized for that access
        403:
          description: User not permitted for that access
        404:
          description: No users retrieved
...
  /users/create/user:
    post:
      summary: Creates user with user role given username, password and role
      tags:
        - Users
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
            example:
              {
                username: user1,
                password: 123,
                role    : user
              }
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                properties:
                  UserID:
                    type: integer
                  role:
                    type: string
                  Username:
                    type: string
                  UserPassword:
                    type: string
                  Created:
                    type: string
              example: 
                  {
                    UserID       : 1,
                    role         : "user",
                    Username     : "user1",
                    UserPassword : "123",
                    Created      : "2024-06-02T19:29:50.000Z"
                  }
        304:
          description: User with that username already exists
        400:
          description: Invalid request Body
          content:
            application/json:
              schema:
                properties:
                  path:
                    type: array
                  property:
                    type: string
                  message:
                    type: string
                  schema:
                    type: string
                  instance:
                    type: object
                  name:
                    type: string
                  argument:
                    type: string
                  stack:
                    type: string
              example: 
                  {
                    path     : [],
                    property : "instance",
                    message  : "requires property \"username\"",
                    schema   : "/users",
                    instance : {},
                    name     : "required",
                    argument : "username",
                    stack    : "instance requires property \"username\""
                  }
        401:
          description: User not authorized for that access
        403:
          description: User not permitted for that access
  /users/create/admin:
    post:
      summary: Creates user with any role given username, password and role
      tags:
        - Users
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
            example:
              {
                username: user1,
                password: 123,
                role    : admin
              }
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                properties:
                  UserID:
                    type: integer
                  role:
                    type: string
                  Username:
                    type: string
                  UserPassword:
                    type: string
                  Created:
                    type: string
              example: 
                  {
                    UserID       : 1,
                    role         : "user",
                    Username     : "user1",
                    UserPassword : "123",
                    Created      : "2024-06-02T19:29:50.000Z"
                  }
        304:
          description: User with that username already exists
        400:
          description: Invalid request Body
          content:
            application/json:
              schema:
                properties:
                  path:
                    type: array
                  property:
                    type: string
                  message:
                    type: string
                  schema:
                    type: string
                  instance:
                    type: object
                  name:
                    type: string
                  argument:
                    type: string
                  stack:
                    type: string
              example: 
                  {
                    path     : [],
                    property : "instance",
                    message  : "requires property \"username\"",
                    schema   : "/users",
                    instance : {},
                    name     : "required",
                    argument : "username",
                    stack    : "instance requires property \"username\""
                  }
        401:
          description: User not authorized for that access
        403:
          description: User not permitted for that access
  /users/:user_id:
    delete:
      summary: Delete user
      tags:
        - Users
      security:
        - JWT: []
      responses:
        200:
          description: not permitted to delete your own account
        401:
          description: User not authorized for that access
        403:
          description: User not permitted for that access
        404:
          description: User does not exist
...